Functions to deal with cmd_t struct
The functions within cmd.c manipulate the cmd_t command data structure. Basic functions for allocating and de-allocating it are present as well as functions to start a process running with the program and arguments contained with a cmd_t and update the data structure based on the state of the job.